create boundary cutter
