+++ src/utils/segmenter/references_computer.py
@@ -1,9 +1,9 @@
-from typing import Callable, List
+from typing import List
 
 from shapely.geometry import LineString, Point
 
+from .boundary_cutter import BoundaryCutter
 from .boundary import Boundary
-from .segment import Segment
 
 
 """
@@ -20,107 +20,36 @@ class ReferencesComputer:
     ):
         self.boundaries = boundaries
 
-    def _get_positions(
-        self,
-        get_sort_key: Callable[[Point], float],
-        length: float,
-        points: List[Point]
-    ) -> List[float]:
-        positions: List[float] = []
-        for i, p in enumerate(points):
-            position = get_sort_key(p)
-            if i > 0 and position <= positions[i-1]:
-                position += length
-                assert position > positions[i-1], \
-                    "Expect current position to be " \
-                    "greater than previous position."
-            positions.append(position)
-        return positions
+    def _consider_boundary_for_segments(self, curr_boundary):
+        curr_boundary = self.boundaries[curr_boundary.idx]
+        curr_boundary.segment_idx_to_neighbors = [
+            [(curr_boundary.idx, curr_boundary.segments[i], False)]
+            for i in range(len(curr_boundary.segments))
+        ]
 
-    def _get_iterations(
-        self,
-        line: LineString,
-        get_sort_key: Callable[[Point], float],
-        length: float,
-    ) -> List[Point]:
-        line_points = [Point(c) for c in line.coords]
-        if line.is_ring:
-            line_points = line_points[:-1]
-        first_boundary = [(p, get_sort_key(p)) for p in line_points]
-        second_boundary = [(p, pos + length) for (p, pos) in first_boundary]
-        if line.is_ring:
-            second_boundary.append((line_points[0], 2*length))
-        return first_boundary + second_boundary
+    def _consider_neighbor_for_loop_segments(self, curr_boundary):
+        for o in curr_boundary.ring_intersections:
+            if o <= curr_boundary.idx:
+                continue
+            other_boundary = self.boundaries[o]
 
-    def _get_segments_helper(
-        self,
-        line: LineString,
-        get_sort_key: Callable[[Point], float],
-        length: float,
-        cutpoints: List[Point]
-    ) -> List[LineString]:
-        positions = self._get_positions(get_sort_key, length, cutpoints)
-        assert len(positions) == len(cutpoints), \
-            "Expect the number of positions to be " \
-            "the same as the number of cutpoints."
-        iterations = self._get_iterations(line, get_sort_key, length)
+            cutpoints_with_end =\
+                curr_boundary.cutpoints + [curr_boundary.cutpoints[0]]
 
-        segments = []
-        segment_coords = None
-        cutpoint_idx = 0
-        for (p, pos) in iterations:
-            if cutpoint_idx == len(cutpoints):
-                break
+            for i in range(len(cutpoints_with_end)-1):
+                start = cutpoints_with_end[i]
+                end = cutpoints_with_end[i+1]
+                segment = curr_boundary.get_segment(start, end)
 
-            if pos < positions[cutpoint_idx]:
-                if segment_coords is None:
-                    continue
-                else:
-                    segment_coords.append(p)
-            else:
-                if segment_coords is None:
-                    segment_coords = []
-                while cutpoint_idx < len(cutpoints) \
-                        and pos >= positions[cutpoint_idx]:
-                    segment_coords.append(cutpoints[cutpoint_idx])
-                    if cutpoint_idx > 0:
-                        segments.append(LineString(segment_coords))
-                        segment_coords = [cutpoints[cutpoint_idx]]
-                        if pos > positions[cutpoint_idx] \
-                                and pos < positions[cutpoint_idx+1]:
-                            segment_coords.append(p)
-                    cutpoint_idx += 1
-
-        return segments
-
-    def _get_segments(
-        self,
-        boundary: Boundary,
-        cutpoints: List[Point],
-    ) -> List[LineString]:
-        segments = self._get_segments_helper(
-            boundary.line,
-            boundary.get_point_sort_key,
-            boundary.line.length, cutpoints,
-        )
-        assert len(segments) == len(cutpoints) - 1, \
-            "Expect number of segments to be one " \
-            "less than number of inputted cutpoints."
-        return segments
-
-    def _compute_segments_per_boundary(self):
-        for b in range(len(self.boundaries)):
-            boundary = self.boundaries[b]
-
-            cutpoints_with_end = boundary.cutpoints + [boundary.cutpoints[0]]
-            segment_lines = self._get_segments(boundary, cutpoints_with_end)
-            boundary.segments = [Segment(boundary, sl) for sl in segment_lines]
-            assert len(boundary.cutpoints) == len(boundary.segments), \
-                "Expect number of segments " \
-                "to equal number of cutpoints."
-            boundary.segment_map = {
-                (s.start, s.end): i for i, s in enumerate(boundary.segments)
-            }
+                other_seg_idx, reverse = \
+                    other_boundary.get_segment_idx_and_reverse(
+                        start, end, segment.line
+                    )
+                other_boundary\
+                    .segment_idx_to_neighbors[
+                        other_seg_idx
+                    ] \
+                    .append((curr_boundary.idx, segment, reverse))
 
     # Get cutpoints to split intersection by.
     def _get_relevant_cutpoints(
@@ -131,73 +60,61 @@ class ReferencesComputer:
         start = Point(intersection.coords[0])
         end = Point(intersection.coords[-1])
 
-        super_line = LineString(boundary.cutpoints)
-        super_segments = self._get_segments_helper(
-            super_line,
-            boundary.get_point_sort_key,
-            boundary.line.length,
+        cutpoints_with_end = boundary.cutpoints + [boundary.cutpoints[0]]
+        boundary_with_just_cutpoints = \
+            Boundary(-1, LineString(cutpoints_with_end))
+        boundary_cutter = BoundaryCutter(
+            boundary_with_just_cutpoints,
             [start, end],
         )
+        super_segments = boundary_cutter.cut_boundary()
         super_segment = super_segments[0]
 
         relevant_cutpoints = [Point(c) for c in super_segment.coords]
         return relevant_cutpoints
 
-    def _compute_boundaries_per_segment(self):
-        for b in range(len(self.boundaries)):
-            curr_boundary = self.boundaries[b]
-            curr_boundary.segment_idx_to_neighbors = [
-                [(b, curr_boundary.segments[i], False)]
-                for i in range(len(curr_boundary.segments))
-            ]
+    def _consider_neighbor_for_line_segments(self, curr_boundary):
+        for o, intersection_segments \
+                in curr_boundary.intersections.items():
+            if o <= curr_boundary.idx:
+                continue  # handled already
+            other_boundary = self.boundaries[o]
 
-        for b in range(len(self.boundaries)):
-            curr_boundary = self.boundaries[b]
+            for intersection_segment in intersection_segments:
+                rel_cutpoints = self._get_relevant_cutpoints(
+                    curr_boundary,
+                    intersection_segment,
+                )
 
-            for n in curr_boundary.ring_intersections:
-                if n <= b:
-                    continue
-                other_boundary = self.boundaries[n]
-                cutpoints_with_end =\
-                    curr_boundary.cutpoints + [curr_boundary.cutpoints[0]]
-                for i in range(len(cutpoints_with_end)-1):
-                    start = cutpoints_with_end[i]
-                    end = cutpoints_with_end[i+1]
+                for i in range(len(rel_cutpoints)-1):
+                    start = rel_cutpoints[i]
+                    end = rel_cutpoints[i+1]
                     segment = curr_boundary.get_segment(start, end)
 
                     other_seg_idx, reverse = \
                         other_boundary.get_segment_idx_and_reverse(
                             start, end, segment.line
                         )
-                    other_boundary.segment_idx_to_neighbors[other_seg_idx] \
-                        .append((b, segment, reverse))
+                    other_boundary \
+                        .segment_idx_to_neighbors[
+                            other_seg_idx
+                        ] \
+                        .append((curr_boundary.idx, segment, reverse))
 
-            for n, intersection_segments \
-                    in curr_boundary.intersections.items():
-                if n <= b:
-                    continue  # handled already
-                other_boundary = self.boundaries[n]
+    def _compute_reference_options_per_segment(self):
+        for b in range(len(self.boundaries)):
+            curr_boundary = self.boundaries[b]
+            self._consider_boundary_for_segments(curr_boundary)
 
-                for intersection_segment in intersection_segments:
-                    rel_cutpoints = self._get_relevant_cutpoints(
-                        curr_boundary,
-                        intersection_segment,
-                    )
+        for b in range(len(self.boundaries)):
+            curr_boundary = self.boundaries[b]
+            self._consider_neighbor_for_loop_segments(curr_boundary)
 
-                    for i in range(len(rel_cutpoints)-1):
-                        start = rel_cutpoints[i]
-                        end = rel_cutpoints[i+1]
-                        segment = curr_boundary.get_segment(start, end)
+        for b in range(len(self.boundaries)):
+            curr_boundary = self.boundaries[b]
+            self._consider_neighbor_for_line_segments(curr_boundary)
 
-                        other_seg_idx, reverse = \
-                            other_boundary.get_segment_idx_and_reverse(
-                                start, end, segment.line
-                            )
-                        other_boundary \
-                            .segment_idx_to_neighbors[other_seg_idx] \
-                            .append((b, segment, reverse))
-
-    def _compute_reference_per_segment(self):
+    def _choose_reference_option_per_segment(self):
         for b in range(len(self.boundaries)):
             boundary = self.boundaries[b]
 
@@ -207,6 +124,5 @@ class ReferencesComputer:
                 boundary.segments[i].set_reference(reference_segment, reverse)
 
     def compute_references(self):
-        self._compute_segments_per_boundary()
-        self._compute_boundaries_per_segment()
-        self._compute_reference_per_segment()
+        self._compute_reference_options_per_segment()
+        self._choose_reference_option_per_segment()
